* Added the delete functionality to the airline maintenance * Added menu for the airline maintenance.

+++ FlightSearch.Business/Components/Maintenance/AirlineComponent.cs
@@ -11,6 +11,7 @@ namespace FlightSearch.Business.Components.Maintenance
     public class AirlineComponent
     {
         private static readonly Logger _log = LogManager.GetLogger("DebugLoggerRules");
+        private const string DEFAULT_LOGO_URL = "https://cdn.shopify.com/s/files/1/0625/8691/3005/products/Pikachu-Knocked-Gloss_01_Top-View_800x.png";
         private AirlineWork _airlineWork = new AirlineWork();
 
         public void CreateNewArline(AirlineEntity airline)
@@ -18,17 +19,29 @@ namespace FlightSearch.Business.Components.Maintenance
             Airlines data = GenericHelper.BusinessMapper.Map<Airlines>(airline);
             data.IsActive = true;
 
+            CheckLogoUrl(airline.LogoUrl, data);
+
             _airlineWork.AirlineRepo.Insert(data);
             _airlineWork.Save();
 
             _log.Info("New airline created");
         }
 
-        public void UpdateAirline(AirlineEntity airline) 
+        private void CheckLogoUrl(string logoUrl, Airlines data)
+        {
+            if (string.IsNullOrEmpty(logoUrl))
+            {
+                data.LogoUrl = DEFAULT_LOGO_URL;
+            }
+        }
+
+        public void UpdateAirline(AirlineEntity airline)
         {
             var data = _airlineWork.AirlineRepo.GetByID(airline.ID);
 
             GenericHelper.BusinessMapper.Map(airline, data);
+            
+            CheckLogoUrl(airline.LogoUrl, data);
 
             _airlineWork.AirlineRepo.Update(data);
 
@@ -37,9 +50,19 @@ namespace FlightSearch.Business.Components.Maintenance
             _log.Info($"Updated airline ID {data.ID}");
         }
 
-        public void DeleteAirline(int id)
+        public void DeleteAirline(int id, bool isFullDelete = false)
         {
-            _airlineWork.AirlineRepo.DeleteByID(id);
+            if (isFullDelete)
+            {
+                _airlineWork.AirlineRepo.DeleteByID(id);
+            }
+            else
+            {
+                var data = _airlineWork.AirlineRepo.GetByID(id);
+                data.IsActive = false;
+                _airlineWork.AirlineRepo.Update(data);
+            }
+
             _airlineWork.Save();
 
             _log.Info($"Deleted airline ID {id}");
@@ -54,7 +77,7 @@ namespace FlightSearch.Business.Components.Maintenance
 
         public List<AirlineEntity> Get()
         {
-            var result = _airlineWork.AirlineRepo.Get();
+            var result = _airlineWork.AirlineRepo.Get(air => air.IsActive);
 
             return result
                 .Select(air => GenericHelper.BusinessMapper.Map<AirlineEntity>(air))
